Rename files in common/ClickableLogo
