+++ simplq/src/components/common/ClickableLogo/ClickableLogo.module.scss
